# Conflicts:
#	app-common/src/main/scala/com/github/ldaniels528/trifecta/AppConstants.scala
#	build.sbt
#	src/main/scala/com/github/ldaniels528/trifecta/io/kafka/KafkaMicroConsumer.scala

+++ app-js/src/main/scala/com/github/ldaniels528/trifecta/sjs/models/QueryRow.scala
@@ -9,7 +9,7 @@ import scala.scalajs.js
 @js.native
 trait QueryRow extends js.Object {
   var topic: js.UndefOr[String] = js.native
-  var __offset: js.UndefOr[Int] = js.native
+  var __offset: js.UndefOr[Double] = js.native
   var __partition: js.UndefOr[Int] = js.native
   var __error: js.UndefOr[String] = js.native
   // and other columns ...
